backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Werkzeug==2.3.7

pymongo==4.6.0
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0

PyPDF2==3.0.1
python-docx==0.8.11

google-generativeai==0.3.0
groq==0.4.2

# AI / ML
sentence-transformers==2.2.2
huggingface-hub==0.16.4
scikit-learn==1.3.2
numpy==1.26.4
faiss-cpu

# Others
bcrypt==4.1.1
langchain==0.1.4
pydantic==2.5.0
aiohttp==3.9.1

# Testing
pytest==7.4.0
pytest-flask==1.2.0